More broken now.
Arduino library supplied by the manufacturer seems to suggest that you clear
nSSEL when sending the instructions to the codec (e.g. ATV) and then set it when
the command is complete. This fills the codec's FIFO until the whole command is
ready. You then toggle nSSEL on each read from the codec in order for it to
shift its data through the FIFO.

As a result I have changed the SPI settings back to manual. And things appear to
be broken...

Will probably take a few days off from this now. Will wait for new, V3 radios to
arrive.